pyproject.toml
[tool.poetry]
name = "goit-web-hw-11"
version = "0.1.0"
description = ""
authors = ["R.Anatolii <tv.weplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-decouple = "^3.8"
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
asyncpg = "^0.30.0"
pydantic = {extras = ["email"], version = "^2.9.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
